Redo version with processes

+++ philo-processes/src/destructor.c
@@ -1,55 +1,24 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   destructors.c                                      :+:      :+:    :+:   */
+/*   destructor.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ipersids <ipersids@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/02/25 16:54:47 by ipersids          #+#    #+#             */
-/*   Updated: 2025/02/25 22:21:01 by ipersids         ###   ########.fr       */
+/*   Created: 2025/03/05 08:38:28 by ipersids          #+#    #+#             */
+/*   Updated: 2025/03/05 14:45:12 by ipersids         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "sophist.h"
+#include "philosophers.h"
 
 /* --------------------- Private function prototypes ----------------------- */
 
-/**
- * @brief Retrieves the error message corresponding to the given exit code.
- * 
- * @param exit_code The exit status code.
- * @return A pointer to the error message string.
- */
 static const char	*get_error_message(int exit_code);
 
 /* --------------------------- Public Functions ---------------------------- */
 
-void	philo_exit_destroy(int exit_code, t_philo *philo)
-{
-	if (philo->processes)
-	{
-		philo_kill(philo);
-		free(philo->processes);
-	}
-	if (SEM_FAILED != philo->sem_lock)
-	{
-		sem_unlink(SEM_LOCK_NAME);
-		sem_close(philo->sem_lock);
-	}
-	if (SEM_FAILED != philo->sem_fork)
-	{
-		sem_unlink(SEM_FORK_NAME);
-		sem_close(philo->sem_fork);
-	}
-	if (SEM_FAILED != philo->sem_error)
-	{
-		sem_unlink(SEM_ERROR_NAME);
-		sem_close(philo->sem_error);
-	}
-	philo_exit(exit_code);
-}
-
-void	philo_exit(int exit_code)
+void	ph_exit(int exit_code)
 {
 	const char	*msg;
 	int			len;
@@ -64,23 +33,42 @@ void	philo_exit(int exit_code)
 	exit(EXIT_FAILURE);
 }
 
-void	philo_kill(t_philo *philo)
+void	ph_destroy_and_exit(int exit_code, t_philo *philo)
 {
-	unsigned int	i;
-	pid_t			*forks;
+	if (philo->processes)
+	{
+		ph_kill(philo);
+		free(philo->processes);
+	}
+	if (0 != philo->print_lock && SEM_FAILED != philo->print_lock)
+		sem_close(philo->print_lock);
+	if (0 != philo->time_lock && SEM_FAILED != philo->time_lock)
+		sem_close(philo->time_lock);
+	if (0 != philo->fork_lock && SEM_FAILED != philo->fork_lock)
+		sem_close(philo->fork_lock);
+	if (0 != philo->meals_lock && SEM_FAILED != philo->meals_lock)
+		sem_close(philo->meals_lock);
+	sem_unlink(SEM_PRINT_NAME);
+	sem_unlink(SEM_TIME_NAME);
+	sem_unlink(SEM_FORK_NAME);
+	sem_unlink(SEM_MEALS_NAME);
+	ph_exit(exit_code);
+}
+
+void	ph_kill(t_philo *philo)
+{
+	int		i;
+	pid_t	*forks;
 
 	i = 0;
 	if (!philo || !philo->processes)
 		return ;
 	forks = philo->processes;
-	while (i < philo->info.forks)
+	while (i < philo->info->forks)
 	{
-		if (0 == kill(forks[i], 0))
-		{
-			printf("I,m killing %d\n", forks[i]); /// testing
+		if (forks[i] && 0 == kill(forks[i], 0))
 			kill(forks[i], SIGTERM);
-		}
-		forks[i] = -1;
+		forks[i] = 0;
 		i++;
 	}
 }
@@ -90,17 +78,18 @@ void	philo_kill(t_philo *philo)
 static const char	*get_error_message(int exit_code)
 {
 	static const char	list[MAX_ERROR_CODE][100] = {
-		"",
+		"Uknown error\n",
 		"Error: wrong amount of arguments (should be 5 or 6)\n",
 		"Error: invalid argument (check forbidden characters)\n",
 		"Error: invalid argument (should be a positive integer more then 0)\n",
 		"Error: invalid argument (should be an integer no more then INT_MAX)\n",
 		"Error: malloc fails\n",
-		"Error: gettimeofday fails\n",
 		"Error: fork fails\n",
 		"Error: semaphore fails\n",
 		"Error: waitpid fails\n"
 	};
 
+	if (MAX_ERROR_CODE <= exit_code || 0 > exit_code)
+		return (list[0]);
 	return (list[exit_code]);
 }
